correção algoritmo de edição de produto

+++ apps/app-web/src/app/data/service/ProdutoService.ts
@@ -21,7 +21,6 @@ import { Produto } from 'libs/data/src/lib/classes';
 export class ProdutoService {
     constructor(private http: HttpClient,
     private AuthenticationService: AuthenticationService,
-    private servicoCategoria: CategoriaService,
     private servicoImagem: ImagemService) { }
 
     Ler(): Observable<entities.Produto[]> {
@@ -65,6 +64,7 @@ export class ProdutoService {
         this.Filtrar(id).subscribe(async Produto =>{
           for(let i =0; i<= Produto[0].Imagem.length ; i++){
             try{
+              if(Produto[0].Imagem[i])
               await this.servicoImagem.deleteImage(Produto[0].Imagem[i]);
             }catch(EX){ console.log(EX); continue;}
           }
